+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Current
-* Use same Main Area pattern at all pages (#355)
+* FR1: Use same Main Area pattern at all pages (#355)
 
 ### Version 3.0.6 (18/06/2021)
 Updates