+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Current (3.0.7)
+* FR3: Image at error 503 refers to old yals.eu hostname (#373)
 * FR2: Submit button no longer disabled on empty input to prevent flaky tests (#367)
 * FR1: YalseeLayout: common layout for all pages (#355)